Fix failing backup rake task when "public/uploads" directory is missing
Forum discussion:
https://groups.google.com/forum/?fromgroups=#!topic/gitlabhq/So4mYIRFsX8

+++ lib/backup/uploads.rb
@@ -11,7 +11,10 @@ module Backup
     # Copy uploads from public/uploads to backup/uploads
     def dump
       FileUtils.mkdir_p(backup_uploads_dir)
-      FileUtils.cp_r(app_uploads_dir, backup_dir)
+      
+      if File.exists?(app_uploads_dir)
+        FileUtils.cp_r(app_uploads_dir, backup_dir)
+      end
     end
 
     def restore
